requirements.txt
absl-py==1.2.0
astor==0.8.1
astunparse==1.6.3
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
cloudpickle==1.2.2
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
dm-reverb==0.8.0
dm-tree==0.1.7
fasteners==0.18
fonttools==4.38.0
future==0.18.2
gast==0.5.3
glfw==2.5.5
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
gym==0.15.3
gym-notices==0.0.8
h5py==3.7.0
idna==3.3
imageio==2.24.0
importlib-metadata==6.0.0
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mock==5.0.1
mpi4py==3.0.2
mujoco-py==2.1.2.14
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
Pillow==9.2.0
portpicker==1.5.2
protobuf==3.20.0
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
pyglet==1.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
requests-oauthlib==1.3.1
rlds==0.1.5
rsa==4.9
scipy==1.7.3
seaborn==0.8.1
six==1.16.0
tensorboard==1.13.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==2.2.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
typing_extensions==4.4.0
urllib3==1.26.11
Werkzeug==2.2.2
xmltodict==0.12.0
zipp==3.11.0
